pyproject.toml
[tool.poetry]
name = "musical-instrument-manual-qa"
version = "0.1.0"
description = "RAG-powered Q&A system for musical instrument manuals"
authors = ["User <user@example.com>"]
readme = "README.md"
packages = [{include = "app", from = "backend"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<3.13"
# Web frameworks
streamlit = ">=1.28.0"
fastapi = ">=0.104.0"
uvicorn = {extras = ["standard"], version = ">=0.24.0"}
pydantic = ">=2.0.0"
pydantic-settings = ">=2.0.0"
python-multipart = ">=0.0.6"
langchain = ">=0.1.0"
langchain-openai = ">=0.0.5"
langchain-community = ">=0.0.20"
chromadb = ">=0.4.0"
pypdf2 = ">=3.0.0"
pymupdf = ">=1.23.0"
pytesseract = ">=0.3.10"
pillow = ">=10.0.0"
python-dotenv = ">=1.0.0"
openai = ">=1.0.0"
tiktoken = ">=0.5.0"
sentence-transformers = ">=2.2.0"
numpy = ">=1.24.0"
pandas = ">=2.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.0.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
mypy = "^1.5.0"

[tool.poetry.scripts]
manual-qa = "app:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true